+++ analyzers/ChainAbuse/requirements.txt
